+++ python/sgl_jax/srt/kernels/gmm/megablox_gmm_backend.py
@@ -1,7 +1,7 @@
 import jax
 import jax.numpy as jnp
 
-from sgl_jax.srt.layers.gmm.megablox_gmm_kernel.gmm import gmm as gmm_kernel
+from sgl_jax.srt.kernels.gmm.megablox_gmm_kernel.gmm import gmm as gmm_kernel
 
 gmm = jax.custom_vjp(gmm_kernel, nondiff_argnums=(3, 4, 7, 8))
