+++ python/sgl_jax/srt/kernels/ragged_paged_attention/tuned_block_sizes.py
@@ -2,7 +2,7 @@
 
 import jax.numpy as jnp
 
-from sgl_jax.srt.layers.attention.flash_attn_kernel.util import (
+from sgl_jax.srt.kernels.ragged_paged_attention.util import (
     get_device_name,
     get_tpu_version,
     next_power_of_2,
