+++ python/sgl_jax/srt/kernels/update_kv_cache/tuned_block_sizes.py
@@ -1,0 +1,273 @@
+best_num_slices_per_block_config = {
+    "hn_8_mcl_80000_nvl_1024_hd_128_ps_64": 16,
+    "hn_8_mcl_80000_nvl_2048_hd_128_ps_64": 32,
+    "hn_8_mcl_80000_nvl_4096_hd_128_ps_64": 64,
+    "hn_8_mcl_80000_nvl_9182_hd_128_ps_64": 32,
+    "hn_8_mcl_80000_nvl_16384_hd_128_ps_64": 512,
+    "hn_8_mcl_160000_nvl_1024_hd_128_ps_64": 16,
+    "hn_8_mcl_160000_nvl_2048_hd_128_ps_64": 8,
+    "hn_8_mcl_160000_nvl_4096_hd_128_ps_64": 64,
+    "hn_8_mcl_160000_nvl_9182_hd_128_ps_64": 32,
+    "hn_8_mcl_160000_nvl_16384_hd_128_ps_64": 128,
+    "hn_8_mcl_320000_nvl_1024_hd_128_ps_64": 16,
+    "hn_8_mcl_320000_nvl_2048_hd_128_ps_64": 16,
+    "hn_8_mcl_320000_nvl_4096_hd_128_ps_64": 32,
+    "hn_8_mcl_320000_nvl_9182_hd_128_ps_64": 16,
+    "hn_8_mcl_320000_nvl_16384_hd_128_ps_64": 256,
+    "hn_8_mcl_640000_nvl_1024_hd_128_ps_64": 4,
+    "hn_8_mcl_640000_nvl_2048_hd_128_ps_64": 64,
+    "hn_8_mcl_640000_nvl_4096_hd_128_ps_64": 64,
+    "hn_8_mcl_640000_nvl_9182_hd_128_ps_64": 32,
+    "hn_8_mcl_640000_nvl_16384_hd_128_ps_64": 64,
+    "hn_8_mcl_1280000_nvl_1024_hd_128_ps_64": 1024,
+    "hn_8_mcl_1280000_nvl_2048_hd_128_ps_64": 4,
+    "hn_8_mcl_1280000_nvl_4096_hd_128_ps_64": 16,
+    "hn_8_mcl_1280000_nvl_9182_hd_128_ps_64": 256,
+    "hn_8_mcl_1280000_nvl_16384_hd_128_ps_64": 1024,
+    "hn_16_mcl_80000_nvl_1024_hd_128_ps_64": 16,
+    "hn_16_mcl_80000_nvl_2048_hd_128_ps_64": 32,
+    "hn_16_mcl_80000_nvl_4096_hd_128_ps_64": 64,
+    "hn_16_mcl_80000_nvl_9182_hd_128_ps_64": 32,
+    "hn_16_mcl_80000_nvl_16384_hd_128_ps_64": 2048,
+    "hn_16_mcl_160000_nvl_1024_hd_128_ps_64": 4,
+    "hn_16_mcl_160000_nvl_2048_hd_128_ps_64": 16,
+    "hn_16_mcl_160000_nvl_4096_hd_128_ps_64": 512,
+    "hn_16_mcl_160000_nvl_9182_hd_128_ps_64": 128,
+    "hn_16_mcl_160000_nvl_16384_hd_128_ps_64": 64,
+    "hn_16_mcl_320000_nvl_1024_hd_128_ps_64": 8,
+    "hn_16_mcl_320000_nvl_2048_hd_128_ps_64": 16,
+    "hn_16_mcl_320000_nvl_4096_hd_128_ps_64": 16,
+    "hn_16_mcl_320000_nvl_9182_hd_128_ps_64": 8,
+    "hn_16_mcl_320000_nvl_16384_hd_128_ps_64": 256,
+    "hn_16_mcl_640000_nvl_1024_hd_128_ps_64": 128,
+    "hn_16_mcl_640000_nvl_2048_hd_128_ps_64": 4,
+    "hn_16_mcl_640000_nvl_4096_hd_128_ps_64": 32,
+    "hn_16_mcl_640000_nvl_9182_hd_128_ps_64": 64,
+    "hn_16_mcl_640000_nvl_16384_hd_128_ps_64": 512,
+    "hn_16_mcl_1280000_nvl_1024_hd_128_ps_64": 128,
+    "hn_16_mcl_1280000_nvl_2048_hd_128_ps_64": 2,
+    "hn_16_mcl_1280000_nvl_4096_hd_128_ps_64": 2048,
+    "hn_16_mcl_1280000_nvl_9182_hd_128_ps_64": 8,
+    "hn_16_mcl_1280000_nvl_16384_hd_128_ps_64": 8,
+    "hn_32_mcl_80000_nvl_1024_hd_128_ps_64": 256,
+    "hn_32_mcl_80000_nvl_2048_hd_128_ps_64": 32,
+    "hn_32_mcl_80000_nvl_4096_hd_128_ps_64": 16,
+    "hn_32_mcl_80000_nvl_9182_hd_128_ps_64": 512,
+    "hn_32_mcl_80000_nvl_16384_hd_128_ps_64": 4096,
+    "hn_32_mcl_160000_nvl_1024_hd_128_ps_64": 8,
+    "hn_32_mcl_160000_nvl_2048_hd_128_ps_64": 8,
+    "hn_32_mcl_160000_nvl_4096_hd_128_ps_64": 512,
+    "hn_32_mcl_160000_nvl_9182_hd_128_ps_64": 64,
+    "hn_32_mcl_160000_nvl_16384_hd_128_ps_64": 1024,
+    "hn_32_mcl_320000_nvl_1024_hd_128_ps_64": 4,
+    "hn_32_mcl_320000_nvl_2048_hd_128_ps_64": 512,
+    "hn_32_mcl_320000_nvl_4096_hd_128_ps_64": 256,
+    "hn_32_mcl_320000_nvl_9182_hd_128_ps_64": 8,
+    "hn_32_mcl_320000_nvl_16384_hd_128_ps_64": 64,
+    "hn_32_mcl_640000_nvl_1024_hd_128_ps_64": 2048,
+    "hn_32_mcl_640000_nvl_2048_hd_128_ps_64": 8,
+    "hn_32_mcl_640000_nvl_4096_hd_128_ps_64": 256,
+    "hn_32_mcl_640000_nvl_9182_hd_128_ps_64": 256,
+    "hn_32_mcl_640000_nvl_16384_hd_128_ps_64": 16,
+    "hn_32_mcl_1280000_nvl_1024_hd_128_ps_64": 256,
+    "hn_32_mcl_1280000_nvl_2048_hd_128_ps_64": 512,
+    "hn_32_mcl_1280000_nvl_4096_hd_128_ps_64": 4096,
+    "hn_32_mcl_1280000_nvl_9182_hd_128_ps_64": 128,
+    "hn_32_mcl_1280000_nvl_16384_hd_128_ps_64": 1024,
+    "hn_8_mcl_80000_nvl_1024_hd_128_ps_128": 8,
+    "hn_8_mcl_80000_nvl_2048_hd_128_ps_128": 32,
+    "hn_8_mcl_80000_nvl_4096_hd_128_ps_128": 64,
+    "hn_8_mcl_80000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_8_mcl_80000_nvl_16384_hd_128_ps_128": 16,
+    "hn_8_mcl_160000_nvl_1024_hd_128_ps_128": 16,
+    "hn_8_mcl_160000_nvl_2048_hd_128_ps_128": 4,
+    "hn_8_mcl_160000_nvl_4096_hd_128_ps_128": 64,
+    "hn_8_mcl_160000_nvl_9182_hd_128_ps_128": 32,
+    "hn_8_mcl_160000_nvl_16384_hd_128_ps_128": 4096,
+    "hn_8_mcl_320000_nvl_1024_hd_128_ps_128": 32,
+    "hn_8_mcl_320000_nvl_2048_hd_128_ps_128": 8,
+    "hn_8_mcl_320000_nvl_4096_hd_128_ps_128": 16,
+    "hn_8_mcl_320000_nvl_9182_hd_128_ps_128": 32,
+    "hn_8_mcl_320000_nvl_16384_hd_128_ps_128": 16,
+    "hn_8_mcl_640000_nvl_1024_hd_128_ps_128": 8,
+    "hn_8_mcl_640000_nvl_2048_hd_128_ps_128": 16,
+    "hn_8_mcl_640000_nvl_4096_hd_128_ps_128": 16,
+    "hn_8_mcl_640000_nvl_9182_hd_128_ps_128": 16,
+    "hn_8_mcl_640000_nvl_16384_hd_128_ps_128": 4096,
+    "hn_8_mcl_1280000_nvl_1024_hd_128_ps_128": 128,
+    "hn_8_mcl_1280000_nvl_2048_hd_128_ps_128": 32,
+    "hn_8_mcl_1280000_nvl_4096_hd_128_ps_128": 256,
+    "hn_8_mcl_1280000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_8_mcl_1280000_nvl_16384_hd_128_ps_128": 32,
+    "hn_16_mcl_80000_nvl_1024_hd_128_ps_128": 128,
+    "hn_16_mcl_80000_nvl_2048_hd_128_ps_128": 8,
+    "hn_16_mcl_80000_nvl_4096_hd_128_ps_128": 2048,
+    "hn_16_mcl_80000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_16_mcl_80000_nvl_16384_hd_128_ps_128": 256,
+    "hn_16_mcl_160000_nvl_1024_hd_128_ps_128": 8,
+    "hn_16_mcl_160000_nvl_2048_hd_128_ps_128": 1024,
+    "hn_16_mcl_160000_nvl_4096_hd_128_ps_128": 2048,
+    "hn_16_mcl_160000_nvl_9182_hd_128_ps_128": 4096,
+    "hn_16_mcl_160000_nvl_16384_hd_128_ps_128": 16,
+    "hn_16_mcl_320000_nvl_1024_hd_128_ps_128": 8,
+    "hn_16_mcl_320000_nvl_2048_hd_128_ps_128": 128,
+    "hn_16_mcl_320000_nvl_4096_hd_128_ps_128": 256,
+    "hn_16_mcl_320000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_16_mcl_320000_nvl_16384_hd_128_ps_128": 32,
+    "hn_16_mcl_640000_nvl_1024_hd_128_ps_128": 2,
+    "hn_16_mcl_640000_nvl_2048_hd_128_ps_128": 8,
+    "hn_16_mcl_640000_nvl_4096_hd_128_ps_128": 8,
+    "hn_16_mcl_640000_nvl_9182_hd_128_ps_128": 32,
+    "hn_16_mcl_640000_nvl_16384_hd_128_ps_128": 4,
+    "hn_16_mcl_1280000_nvl_1024_hd_128_ps_128": 32,
+    "hn_16_mcl_1280000_nvl_2048_hd_128_ps_128": 16,
+    "hn_16_mcl_1280000_nvl_4096_hd_128_ps_128": 128,
+    "hn_16_mcl_1280000_nvl_9182_hd_128_ps_128": 64,
+    "hn_16_mcl_1280000_nvl_16384_hd_128_ps_128": 2048,
+    "hn_32_mcl_80000_nvl_1024_hd_128_ps_128": 2,
+    "hn_32_mcl_80000_nvl_2048_hd_128_ps_128": 4,
+    "hn_32_mcl_80000_nvl_4096_hd_128_ps_128": 256,
+    "hn_32_mcl_80000_nvl_9182_hd_128_ps_128": 8,
+    "hn_32_mcl_80000_nvl_16384_hd_128_ps_128": 512,
+    "hn_32_mcl_160000_nvl_1024_hd_128_ps_128": 16,
+    "hn_32_mcl_160000_nvl_2048_hd_128_ps_128": 1024,
+    "hn_32_mcl_160000_nvl_4096_hd_128_ps_128": 512,
+    "hn_32_mcl_160000_nvl_9182_hd_128_ps_128": 8,
+    "hn_32_mcl_160000_nvl_16384_hd_128_ps_128": 4096,
+    "hn_32_mcl_320000_nvl_1024_hd_128_ps_128": 256,
+    "hn_32_mcl_320000_nvl_2048_hd_128_ps_128": 256,
+    "hn_32_mcl_320000_nvl_4096_hd_128_ps_128": 256,
+    "hn_32_mcl_320000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_32_mcl_320000_nvl_16384_hd_128_ps_128": 4,
+    "hn_32_mcl_640000_nvl_1024_hd_128_ps_128": 512,
+    "hn_32_mcl_640000_nvl_2048_hd_128_ps_128": 4096,
+    "hn_32_mcl_640000_nvl_4096_hd_128_ps_128": 4,
+    "hn_32_mcl_640000_nvl_9182_hd_128_ps_128": 2,
+    "hn_32_mcl_640000_nvl_16384_hd_128_ps_128": 1024,
+    "hn_32_mcl_1280000_nvl_1024_hd_128_ps_128": 32,
+    "hn_32_mcl_1280000_nvl_2048_hd_128_ps_128": 2048,
+    "hn_32_mcl_1280000_nvl_4096_hd_128_ps_128": 128,
+    "hn_32_mcl_1280000_nvl_9182_hd_128_ps_128": 1024,
+    "hn_32_mcl_1280000_nvl_16384_hd_128_ps_128": 1024,
+    "hn_8_mcl_80000_nvl_1024_hd_128_ps_256": 2,
+    "hn_8_mcl_80000_nvl_2048_hd_128_ps_256": 4,
+    "hn_8_mcl_80000_nvl_4096_hd_128_ps_256": 2,
+    "hn_8_mcl_80000_nvl_9182_hd_128_ps_256": 512,
+    "hn_8_mcl_80000_nvl_16384_hd_128_ps_256": 2048,
+    "hn_8_mcl_160000_nvl_1024_hd_128_ps_256": 4,
+    "hn_8_mcl_160000_nvl_2048_hd_128_ps_256": 256,
+    "hn_8_mcl_160000_nvl_4096_hd_128_ps_256": 128,
+    "hn_8_mcl_160000_nvl_9182_hd_128_ps_256": 2048,
+    "hn_8_mcl_160000_nvl_16384_hd_128_ps_256": 2048,
+    "hn_8_mcl_320000_nvl_1024_hd_128_ps_256": 4,
+    "hn_8_mcl_320000_nvl_2048_hd_128_ps_256": 1024,
+    "hn_8_mcl_320000_nvl_4096_hd_128_ps_256": 64,
+    "hn_8_mcl_320000_nvl_9182_hd_128_ps_256": 16,
+    "hn_8_mcl_320000_nvl_16384_hd_128_ps_256": 512,
+    "hn_8_mcl_640000_nvl_1024_hd_128_ps_256": 1024,
+    "hn_8_mcl_640000_nvl_2048_hd_128_ps_256": 8,
+    "hn_8_mcl_640000_nvl_4096_hd_128_ps_256": 16,
+    "hn_8_mcl_640000_nvl_9182_hd_128_ps_256": 16,
+    "hn_8_mcl_640000_nvl_16384_hd_128_ps_256": 4096,
+    "hn_8_mcl_1280000_nvl_1024_hd_128_ps_256": 64,
+    "hn_8_mcl_1280000_nvl_2048_hd_128_ps_256": 2,
+    "hn_8_mcl_1280000_nvl_4096_hd_128_ps_256": 2048,
+    "hn_8_mcl_1280000_nvl_9182_hd_128_ps_256": 1024,
+    "hn_8_mcl_1280000_nvl_16384_hd_128_ps_256": 128,
+    "hn_16_mcl_80000_nvl_1024_hd_128_ps_256": 2,
+    "hn_16_mcl_80000_nvl_2048_hd_128_ps_256": 16,
+    "hn_16_mcl_80000_nvl_4096_hd_128_ps_256": 64,
+    "hn_16_mcl_80000_nvl_9182_hd_128_ps_256": 256,
+    "hn_16_mcl_80000_nvl_16384_hd_128_ps_256": 16,
+    "hn_16_mcl_160000_nvl_1024_hd_128_ps_256": 4,
+    "hn_16_mcl_160000_nvl_2048_hd_128_ps_256": 2,
+    "hn_16_mcl_160000_nvl_4096_hd_128_ps_256": 128,
+    "hn_16_mcl_160000_nvl_9182_hd_128_ps_256": 16,
+    "hn_16_mcl_160000_nvl_16384_hd_128_ps_256": 8,
+    "hn_16_mcl_320000_nvl_1024_hd_128_ps_256": 16,
+    "hn_16_mcl_320000_nvl_2048_hd_128_ps_256": 8,
+    "hn_16_mcl_320000_nvl_4096_hd_128_ps_256": 4,
+    "hn_16_mcl_320000_nvl_9182_hd_128_ps_256": 8,
+    "hn_16_mcl_320000_nvl_16384_hd_128_ps_256": 8,
+    "hn_16_mcl_640000_nvl_1024_hd_128_ps_256": 512,
+    "hn_16_mcl_640000_nvl_2048_hd_128_ps_256": 1024,
+    "hn_16_mcl_640000_nvl_4096_hd_128_ps_256": 2048,
+    "hn_16_mcl_640000_nvl_9182_hd_128_ps_256": 4096,
+    "hn_16_mcl_640000_nvl_16384_hd_128_ps_256": 32,
+    "hn_16_mcl_1280000_nvl_1024_hd_128_ps_256": 4,
+    "hn_16_mcl_1280000_nvl_2048_hd_128_ps_256": 2,
+    "hn_16_mcl_1280000_nvl_4096_hd_128_ps_256": 1024,
+    "hn_16_mcl_1280000_nvl_9182_hd_128_ps_256": 2048,
+    "hn_16_mcl_1280000_nvl_16384_hd_128_ps_256": 16,
+    "hn_32_mcl_80000_nvl_1024_hd_128_ps_256": 4,
+    "hn_32_mcl_80000_nvl_2048_hd_128_ps_256": 256,
+    "hn_32_mcl_80000_nvl_4096_hd_128_ps_256": 4096,
+    "hn_32_mcl_80000_nvl_9182_hd_128_ps_256": 128,
+    "hn_32_mcl_80000_nvl_16384_hd_128_ps_256": 512,
+    "hn_32_mcl_160000_nvl_1024_hd_128_ps_256": 64,
+    "hn_32_mcl_160000_nvl_2048_hd_128_ps_256": 4096,
+    "hn_32_mcl_160000_nvl_4096_hd_128_ps_256": 4096,
+    "hn_32_mcl_160000_nvl_9182_hd_128_ps_256": 256,
+    "hn_32_mcl_160000_nvl_16384_hd_128_ps_256": 128,
+    "hn_32_mcl_320000_nvl_1024_hd_128_ps_256": 4,
+    "hn_32_mcl_320000_nvl_2048_hd_128_ps_256": 64,
+    "hn_32_mcl_320000_nvl_4096_hd_128_ps_256": 1024,
+    "hn_32_mcl_320000_nvl_9182_hd_128_ps_256": 256,
+    "hn_32_mcl_320000_nvl_16384_hd_128_ps_256": 32,
+    "hn_32_mcl_640000_nvl_1024_hd_128_ps_256": 256,
+    "hn_32_mcl_640000_nvl_2048_hd_128_ps_256": 8,
+    "hn_32_mcl_640000_nvl_4096_hd_128_ps_256": 64,
+    "hn_32_mcl_640000_nvl_9182_hd_128_ps_256": 32,
+    "hn_32_mcl_640000_nvl_16384_hd_128_ps_256": 32,
+    "hn_32_mcl_1280000_nvl_1024_hd_128_ps_256": 256,
+    "hn_32_mcl_1280000_nvl_2048_hd_128_ps_256": 2048,
+    "hn_32_mcl_1280000_nvl_4096_hd_128_ps_256": 8,
+    "hn_32_mcl_1280000_nvl_9182_hd_128_ps_256": 4,
+    "hn_32_mcl_1280000_nvl_16384_hd_128_ps_256": 2,
+}
+
+
+def get_best_num_slices_per_block(head_num, cache_len, new_kv_len, head_dim, page_size):
+    # keep same to original implementation
+    num_slices_per_block = 4 if page_size == 1 else page_size
+
+    return num_slices_per_block
+
+    # note: the following logic will be supported in the future, the best num_slices_per_block is right tested well currently.
+    # search domain, ensure list is sorted
+    head_num_config = [8, 16, 32]
+    max_cache_len_config = [80000, 160000, 320000, 640000, 1280000]
+    new_kv_len_config = [1024, 2048, 4096, 9182, 16384]
+    head_dim_config = [128]
+    page_size_config = [64, 128, 256]
+
+    def find_value(lst, target_num) -> int:
+        left, right = 0, len(lst) - 1
+
+        if not lst or target_num < lst[0] or target_num > lst[-1]:
+            return -1
+
+        while left <= right:
+            mid = (left + right) // 2
+            if lst[mid] == target_num:
+                return lst[mid]
+            elif lst[mid] < target_num:
+                left = mid + 1
+            else:
+                right = mid - 1
+
+        if left < len(lst):
+            return lst[left]
+        else:
+            return -1
+
+    hn_val = find_value(head_num_config, head_num)
+    mcl_val = find_value(max_cache_len_config, cache_len)
+    nkl_val = find_value(new_kv_len_config, new_kv_len)
+    hd_val = find_value(head_dim_config, head_dim)
+    ps_val = find_value(page_size_config, page_size)
+
+    if hn_val != -1 and mcl_val != -1 and nkl_val != -1 and hd_val != -1 and ps_val != -1:
+        return best_num_slices_per_block_config[
+            f"hn_{hn_val}_mcl_{mcl_val}_nvl_{nkl_val}_hd_{hd_val}_ps_{ps_val}"
+        ]
